requirements.txt
# Core dependencies
google-cloud-aiplatform==1.36.4
vertexai==0.0.1  # Vertex AI Agent SDK

# Document processing
python-magic==0.4.27
python-docx==1.0.1
PyPDF2==3.0.1

# Deployment
pyyaml==6.0.1